refactor: move base input and max length to a new folder
