+++ features/core/components/form/ui/index.ts
@@ -1,0 +1,5 @@
+import { BaseInput } from "./BaseInput";
+import { BaseTextArea } from "./BaseTextArea";
+import { MaxLength } from "./MaxLength";
+
+export { BaseInput, BaseTextArea, MaxLength };
